app/requirements.txt
# FastAPI + Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# AI APIs
anthropic==0.18.1
openai==1.12.0

# PDF Handling
pypdf==4.0.1
fillpdf==0.7.2
reportlab==4.1.0

# Email
resend==0.8.0

# Utilities
python-dotenv==1.0.1
pydantic==2.6.0
pydantic-settings==2.1.0
httpx==0.26.0

# Testing
pytest==8.0.0
pytest-asyncio==0.23.4
pytest-cov==4.1.0
pytest-mock==3.12.0
respx==0.20.2

# Linting & Formatting
ruff==0.2.0
